useChannelMetadata uses creation info if no updated metadata
