+++ packages/use-arcade/src/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './useArcadeRelay'
+export * from './useChannelsCreated'
 export * from './useChannelMessages'
 export * from './useLastChannelMessage'
