+++ packages/use-arcade/src/index.ts
@@ -3,17 +3,9 @@ export { ArcadeContext } from './context'
 export { useArcadeRelay, UseArcadeRelayActions } from './hooks/useArcadeRelay'
 export { useBoolean } from './hooks/useBoolean'
 
-export { useChannelMessages, useLastChannelMessage } from './hooks'
+export { useChannelMessages, useChannelsCreated, useLastChannelMessage } from './hooks'
 
-export {
-  Channel,
-  Message,
-  setActiveChannelId,
-  store,
-  useActiveChannelId,
-  useChannelsCreated,
-  useRideRequests,
-} from './store'
+export { Channel, Message, setActiveChannelId, store, useActiveChannelId, useRideRequests } from './store'
 
 export type { UseBoolean, UseBooleanActions } from './hooks/useBoolean'
