+++ packages/use-arcade/src/store/index.ts
@@ -30,22 +30,6 @@ export const useActiveChannelId: () => string | null = () => {
   return snapshot.activeChannelId
 }
 
-export const useChannelsCreated: () => Channel[] = () => {
-  const snapshot = useSnapshot(store)
-  const eventArray = Array.from(snapshot.events.values()) as NostrEvent[]
-  return eventArray
-    .filter((event: NostrEvent) => event.kind === NostrKind.channelcreate)
-    .map((event: NostrEvent) => {
-      const parsedEvent = JSON.parse(event.content)
-      const cleanedEvent = Object.assign({}, event)
-      delete cleanedEvent.content
-      return {
-        ...cleanedEvent,
-        ...parsedEvent,
-      } as Channel
-    })
-}
-
 export const useRideRequests = () => {
   const snapshot = useSnapshot(store)
   const eventArray = Array.from(snapshot.events.values()) as NostrEvent[]
